Actualización de variables de entorno y modelos de datos

+++ models/Data_empresa_garantia.js
@@ -12,6 +12,6 @@ numero_garantia:{type: DataTypes.INTEGER,allowNull: false,trim: true},
 fecha_inicio_vigencia:{type: DataTypes.STRING,allowNull: false,trim: true},
 fecha_fin_vigencia:{type: DataTypes.STRING,allowNull: false,trim: true},
 valor_garantia:{type: DataTypes.STRING,allowNull: false,trim: true},
-costo_garantia:{type: DataTypes.INTEGER,allowNull: false,trim: true},
+costo_garantia:{type: DataTypes.FLOAT,allowNull: false,trim: true},
 })
 module.exports = data_empresa_garantiaSchema;
